add: criteria based tricks
there will now be two kinds of tricks. tricks that check for a condition and then run, or tricks that are fail safe and will work everytime. form is a separate component now so as to stop unmounting of ReasonBox component unnecesarily which causes issues with the random trick chosen. this keeps changing as even an input box update will re-render the component thus picking a new trick.